refactor part 5: move testing annotator module to the module providing fake classes
